# solution/backend/pkg/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "PROD Backend 2025 Advertising Platform API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/ai/text/generate": {
            "post": {
                "description": "Создает текст для рекламной кампании по пожеланиям от пользователя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AI"
                ],
                "summary": "Сгенерировать текст для рекламной кампании",
                "parameters": [
                    {
                        "description": "Название товара и пожелания",
                        "name": "GenerateInfo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.GenerateTextRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Обновленная кампания",
                        "schema": {
                            "$ref": "#/definitions/model.GenerateTextResponse"
                        }
                    },
                    "400": {
                        "description": "Невалидные данные",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/model.GenerateTextResponse"
                        }
                    }
                }
            }
        },
        "/ai/text/moderation": {
            "post": {
                "description": "Включает или выключает модерацию текстов и названий рекламных кампаний",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AI"
                ],
                "summary": "Включить или выключить модерацию текстов",
                "parameters": [
                    {
                        "description": "true - включить модерацию, false - выключить модерацию",
                        "name": "Moderation",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ModerationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Обновленная кампания",
                        "schema": {
                            "$ref": "#/definitions/model.ModerationRequest"
                        }
                    },
                    "400": {
                        "description": "Невалидные данные",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.ErrorResponse": {
            "type": "object",
            "required": [
                "message",
                "status"
            ],
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "model.GenerateTextRequest": {
            "type": "object",
            "required": [
                "title",
                "wishes"
            ],
            "properties": {
                "gender": {
                    "type": "string",
                    "enum": [
                        "MALE",
                        "FEMALE"
                    ]
                },
                "location": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "wishes": {
                    "type": "string",
                    "maxLength": 100
                }
            }
        },
        "model.GenerateTextResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "model.ModerationRequest": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "status": {
                    "type": "boolean"
                }
            }
        },
...
    }
}